Laberinto

Description

Módulo para crear y manejar un Laberinto.

Tipos de Datos

data Laberinto

Conocimientos sobre un Laberinto

data Indicador

Indicador de cuál camino relaciona una Trifurcacion y un Laberinto

type Ruta = [Indicador]

Ruta hecha por Indicadores

Funciones de Construcción

laberintoStart :: Laberinto

Función que genera una Trifurcación sin salida

tesoroAdd :: String -> Laberinto -> Laberinto

Crea un tesoro, con su descripción y un Laberinto | indicando qué encontrarán si pasan por alto el tesoro

trifurcacionAdd :: Laberinto -> Laberinto -> Indicador -> Laberinto

Conecta una Trifurcación a un Laberinto siguiendo la dirección de un Indicador

Funciones de Acceso

seguirRuta :: Laberinto -> Ruta -> Laberinto

Dado un Laberinto y una Ruta, devuelve el Laberinto alcanzado al seguir la Ruta

seguirIzq :: Laberinto -> Laberinto

Dado un Laberinto, devuelve el Laberinto alcanzado al seguir a la Izquierda

seguirRect :: Laberinto -> Laberinto

Dado un Laberinto, devuelve el Laberinto alcanzado al seguir Recto

seguirDer :: Laberinto -> Laberinto

Dado un Laberinto, devuelve el Laberinto alcanzado al seguir a la Derecha